@@ -10,41 +10,38 @@
     "node": ">=14"
   },
   "scripts": {
-    "lint": "yarn prettier && yarn eslint",
-    "eslint": "next lint --fix --dir src",
-    "prettier": "prettier -l \"./src/**/*.{js,jsx,md}\"",
+    "lint": "next lint --fix --dir app",
     "dev": "next dev",
     "build": "next build",
-    "export": "EXPORT=true next build && EXPORT=true next export",
     "analyze": "ANALYZE=true next build",
     "start": "next start"
   },
   "dependencies": {
-    "@react-three/drei": "^9.56.5",
-    "@react-three/fiber": "^8.10.1",
+    "@ducanh2912/next-pwa": "^8.2.0",
+    "@react-three/drei": "^9.65.3",
+    "@react-three/fiber": "^8.12.0",
     "glsl-random": "^0.0.5",
-    "next": "^13.1.6",
-    "next-pwa": "^5.6.0",
+    "next": "^13.3.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "three": "^0.149.0",
-    "three-stdlib": "^2.21.6"
+    "three": "^0.151.3",
+    "three-stdlib": "^2.21.8",
+    "tunnel-rat": "^0.1.2"
   },
   "devDependencies": {
-    "@next/bundle-analyzer": "^13.1.6",
-    "autoprefixer": "^10.4.12",
-    "eslint": "^8.33.0",
-    "eslint-config-next": "^13.1.6",
-    "eslint-config-prettier": "^8.6.0",
-    "eslint-plugin-tailwindcss": "^3.9.0",
-
+    "@next/bundle-analyzer": "^13.3.0",
+    "autoprefixer": "^10.4.14",
+    "eslint": "^8.38.0",
+    "eslint-config-next": "^13.3.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-tailwindcss": "^3.11.0",
     "file-loader": "^6.2.0",
     "glslify": "^7.1.1",
     "glslify-loader": "^2.0.0",
     "postcss": "^8.4.21",
-    "prettier": "^2.8.3",
+    "prettier": "^2.8.7",
     "raw-loader": "^4.0.2",
-    "tailwindcss": "^3.2.4",
+    "tailwindcss": "^3.3.1",
     "url-loader": "^4.1.1"
   }
 }
