@@ -1,12 +1,32 @@
+'use client'
+
+import { useGLTF } from '@react-three/drei'
+import { useFrame } from '@react-three/fiber'
 import * as THREE from 'three'
 import { useMemo, useRef, useState } from 'react'
-import { useRouter } from 'next/router'
-import { useFrame } from '@react-three/fiber'
-import { Line, useCursor } from '@react-three/drei'
+import { Line, useCursor, MeshDistortMaterial } from '@react-three/drei'
+import { useRouter } from 'next/navigation'
 
-export default function Logo({ route, ...props }) {
+export const Blob = ({ route = '/', ...props }) => {
   const router = useRouter()
+  const [hovered, hover] = useState(false)
+  useCursor(hovered)
+  return (
+    <mesh
+      onClick={() => router.push(route)}
+      onPointerOver={() => hover(true)}
+      onPointerOut={() => hover(false)}
+      {...props}>
+      <sphereGeometry args={[1, 64, 64]} />
+      <MeshDistortMaterial roughness={0} color={hovered ? 'hotpink' : '#1fb2f5'} />
+    </mesh>
+  )
+}
+
+export const Logo = ({ route = '/blob', ...props }) => {
   const mesh = useRef(null)
+  const router = useRouter()
+
   const [hovered, hover] = useState(false)
   const points = useMemo(() => new THREE.EllipseCurve(0, 0, 3, 1.15, 0, 2 * Math.PI, false, 0).getPoints(100), [])
 
@@ -32,4 +52,17 @@ export default function Logo({ route, ...props }) {
       </mesh>
     </group>
   )
+}
+
+export function Duck(props) {
+  const { scene } = useGLTF('/duck.glb')
+
+  useFrame((state, delta) => (scene.rotation.y += delta))
+
+  return <primitive object={scene} {...props} />
+}
+export function Dog(props) {
+  const { scene } = useGLTF('/dog.glb')
+
+  return <primitive object={scene} {...props} />
 }
