@@ -1,13 +1,12 @@
-import Head from 'next/head'
-
-const titleDefault = 'React Three Next Starter'
+const title = 'React Three Next Starter'
 const url = 'https://react-three-next.vercel.app/'
 const description = 'The easiest and fastest way to create a 3D website using React Three Fiber and NextJS'
 const author = 'Author'
+const twitter = '@pmndrs'
 
-export default function Header({ title = titleDefault }) {
+export default function Head() {
   return (
-    <Head>
+    <>
       {/* Recommended Meta Tags */}
       <meta charSet='utf-8' />
       <meta name='language' content='english' />
@@ -48,14 +47,14 @@ export default function Header({ title = titleDefault }) {
         <meta name="HandheldFriendly" content="true"/>  */}
       <meta name='viewport' content='width=device-width, minimum-scale=1, initial-scale=1.0' />
       <meta name='theme-color' content='#000' />
-      <link rel='shortcut icon' href='/icons/favicon.ico' />
+      <link rel='shortcut icon' href='/icons/apple-touch-icon.png' />
 
       {/* 
       Twitter Summary card
         documentation: https://dev.twitter.com/cards/getting-started
         Be sure validate your Twitter card markup on the documentation site. */}
       <meta name='twitter:card' content='summary' />
-      <meta name='twitter:site' content='@onirenaud' />
-    </Head>
+      <meta name='twitter:site' content={twitter} />
+    </>
   )
 }
