@@ -1,15 +1,16 @@
-import { Canvas } from '@react-three/fiber'
-import { OrbitControls, Preload } from '@react-three/drei'
+'use client'
 
-export default function Scene({ children, ...props }) {
+import { Canvas } from '@react-three/fiber'
+import { Preload } from '@react-three/drei'
+import { r3f } from '@/helpers/global'
+
+export default function Scene({ ...props }) {
   // Everything defined in here will persist between route changes, only children are swapped
   return (
     <Canvas {...props}>
-      <directionalLight intensity={0.75} />
-      <ambientLight intensity={0.75} />
-      {children}
+      {/* @ts-ignore */}
+      <r3f.Out />
       <Preload all />
-      <OrbitControls />
     </Canvas>
   )
 }
